Recfactoring based on PR comments.
