+++ src/software/geom/voronoi_diagram.cpp
@@ -1,16 +1,15 @@
-#include "software/geom/voronoi_util.h"
+#include "software/geom/voronoi_diagram.h"
 
 #include <g3log/g3log.hpp>
 #include <g3log/loglevels.hpp>
 
 #include "software/geom/util.h"
 
-std::vector<Point> findVoronoiEdgeRecIntersects(const VoronoiDiagram &diagram,
-                                                Rectangle bounding_box)
+std::vector<Point> VoronoiDiagram::findVoronoiEdgeRecIntersects(Rectangle bounding_box)
 {
     std::vector<Point> intersects;
 
-    for (auto edge : diagram.getDiagram().edges())
+    for (auto edge : diagram.edges())
     {
         // Edges extending outside the rectangle will be infinite edges
         if (!edge.is_finite() && edge.is_primary())
@@ -34,8 +33,8 @@ std::vector<Point> findVoronoiEdgeRecIntersects(const VoronoiDiagram &diagram,
                 //      |                 v                   |
                 //      |                                     |
                 //      +-------------------------------------+
-                Point p1    = diagram.getPointsUsedForGeneration()[edge.cell()->source_index()];
-                Point p2    = diagram.getPointsUsedForGeneration()[edge.twin()->cell()->source_index()];
+                Point p1 = points[edge.cell()->source_index()];
+                Point p2 = points[edge.twin()->cell()->source_index()];
                 double endX = (p1.y() - p2.y());
                 double endY = (p1.x() - p2.x()) * -1;
                 // Extend the edge out to beyond the rectangle to ensure interception
@@ -53,8 +52,8 @@ std::vector<Point> findVoronoiEdgeRecIntersects(const VoronoiDiagram &diagram,
     return intersects;
 }
 
-std::vector<Circle> voronoiVerticesToOpenCircles(const VoronoiDiagram &diagram,
-                                                 const Rectangle &bounding_box)
+std::vector<Circle> VoronoiDiagram::voronoiVerticesToOpenCircles(
+    const Rectangle &bounding_box)
 {
     // For each vertex, construct it's delauney triangle and then compute the largest
     // empty circle around it
@@ -68,7 +67,7 @@ std::vector<Circle> voronoiVerticesToOpenCircles(const VoronoiDiagram &diagram,
     // https://stackoverflow.com/questions/34342038/how-to-triangulate-polygons-in-boost
 
     std::vector<Circle> empty_circles;
-    for (auto vertex : diagram.getDiagram().vertices())
+    for (auto vertex : diagram.vertices())
     {
         // We only want to consider vertices within our rectangle
         if (bounding_box.containsPoint(Point(vertex.x(), vertex.y())))
@@ -85,7 +84,7 @@ std::vector<Circle> voronoiVerticesToOpenCircles(const VoronoiDiagram &diagram,
                     continue;
                 }
 
-                triangle.push_back(diagram.getPointsUsedForGeneration()[cell->source_index()]);
+                triangle.push_back(points[cell->source_index()]);
                 if (triangle.size() == 3)
                 {
                     // Find the smallest distance from the vertex to a vertex on it's
